Move chart data in useChartData hook
